pyproject.toml
[project]
name = "applypilot"
version = "1.0.0-alpha"
description = "AI-powered end-to-end job application pipeline"
readme = "README.md"
license = "AGPL-3.0-only"
requires-python = ">=3.11"
authors = [{name = "Pickle-Pixel"}]
keywords = ["job-search", "automation", "ai", "resume", "apply", "job-application"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "jsonschema>=4.0",
    "litellm>=1.63.0",
    "httpx>=0.24",
    "tenacity>=8.2",
    "beautifulsoup4>=4.12",
    "playwright>=1.40",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "mcp>=1.9",
    "python-jobspy>=1.1",
    "markdownify>=0.14.1",
    "websocket-client>=1.0",
    "transitions>=0.9",
    "local-geocode>=0.0.2",
    ]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1", "pytest-cov>=4.0"]
bedrock = ["boto3>=1.28"]

[project.scripts]
applypilot = "applypilot.cli:app"

[project.urls]
Homepage = "https://github.com/ibarrajo/ApplyPilot"
Repository = "https://github.com/ibarrajo/ApplyPilot"
Issues = "https://github.com/ibarrajo/ApplyPilot/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/applypilot"]

[tool.hatch.build]
artifacts = ["src/applypilot/config/*.yaml"]

[tool.pytest.ini_options]
addopts = "--cov=src/applypilot --cov-report=term-missing -m 'not live'"
testpaths = ["tests"]
markers = [
    "live: real API calls — run with: pytest -m live",
    "smoke: lightweight smoke tests",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
ignore = ["E402", "F405", "E741"]
